+++ Gateway/Request/APMBuilder.php
@@ -8,6 +8,7 @@ use Omise\Payment\Model\Config\Alipay;
 use Omise\Payment\Model\Config\Installment;
 use Omise\Payment\Model\Config\Internetbanking;
 use Omise\Payment\Model\Config\Tesco;
+use Omise\Payment\Model\Config\Truemoney;
 use Omise\Payment\Observer\InstallmentDataAssignObserver;
 use Omise\Payment\Observer\InternetbankingDataAssignObserver;
 
@@ -28,6 +29,11 @@ class APMBuilder implements BuilderInterface
      * @var string
      */
     const SOURCE_INSTALLMENT_TERMS = 'installment_terms';
+
+    /**
+     * @var string
+     */
+    const SOURCE_PHONE_NUM = 'phone_number';
 
     /**
      * @var string
@@ -80,6 +86,12 @@ class APMBuilder implements BuilderInterface
                 $paymentInfo[self::SOURCE] = [
                     self::SOURCE_TYPE              => $method->getAdditionalInformation(InstallmentDataAssignObserver::OFFSITE),
                     self::SOURCE_INSTALLMENT_TERMS => $method->getAdditionalInformation(InstallmentDataAssignObserver::TERMS),
+                ];
+                break;
+            case Truemoney::CODE:
+                $paymentInfo[self::SOURCE] = [
+                    self::SOURCE_TYPE      => 'truemoney',
+                    self::SOURCE_PHONE_NUM => $method->getOrder()->getBillingAddress()->getTelephone(),
                 ];
                 break;
         }